+++ crates/mirror-mirror/src/foreign_impls/boxed.rs
@@ -1,4 +1,3 @@
-use alloc::borrow::Cow;
 use alloc::boxed::Box;
 use core::any::Any;
 use core::fmt;
@@ -12,7 +11,6 @@ use crate::Reflect;
 use crate::ReflectMut;
 use crate::ReflectOwned;
 use crate::ReflectRef;
-use crate::TypeDescriptor;
 use crate::Value;
 
 impl<T> DescribeType for Box<T>
@@ -28,10 +26,6 @@ impl<T> Reflect for Box<T>
 where
     T: Reflect + DescribeType,
 {
-    fn type_descriptor(&self) -> Cow<'static, TypeDescriptor> {
-        <T as DescribeType>::type_descriptor()
-    }
-
     fn as_any(&self) -> &dyn Any {
         <T as Reflect>::as_any(self)
     }